requirements.txt
email_validator==2.2.0
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Security-Too==5.4.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
pytz==2024.1
python-dotenv==1.0.1
setuptools==70.1.0
